move usefilter hook tests out of ticketstable.test.tsx
